requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
pandas 
numpy
python-dotenv
streamlit
ipykernel
huggingface_hub
astrapy
google-generativeai
langchain_google_genai
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai

sentence-transformers
uvicorn
ctransformers
langchain
python-box
streamlit